requirements.txt
weasyprint==61.0
markdown==3.6
Jinja2==3.1.4
PyYAML==6.0.1
